pyproject.toml
[tool.poetry]
name = "converter"
version = "0.1.0"
description = ""
authors = ["Leonid Sliusar <loslyusar@gmail.com>"]
readme = "README.md"
packages = [
  { include = "src", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.97.0"}
pillow = "^9.5.0"
pytesseract = "^0.3.10"
aiofiles = "^23.1.0"
pydantic = "^1.10.9"
pdf2image = "^1.16.3"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src", "src/form"
]